removed warning by using explicit type conversion

+++ cardcomm/pkcs11/src/cardlayer/pkcs15.cpp
@@ -145,7 +145,7 @@ namespace eIDMW
 
 	unsigned long CPKCS15::PrivKeyCount()
 	{
-		return keymap[m_poCard->GetAppletVersion()].size();
+		return (unsigned long)(keymap[m_poCard->GetAppletVersion()].size());
 	}
 
 	tPrivKey CPKCS15::GetPrivKey(unsigned long ulIndex)
